+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # v 0.8.0 ( unreleased ) 
 ### New Features
 * Added new capabilities to add , modify and delete devices online ( avoids restart the gathering ocess  on the other devices) 
+* Go 1.9.8 to 1.11 binaries 
+* Added a new Conversión parameter to most of the snmpmetric types.
+* Implement HexString to INTEGUER conversion , implements #310
+
+### fixes
+* Fix CVE-2018-3721 vulnerability
+
+### breaking changes
+* IMPORTANT!!!: OCTECSTRING  SnmpMetric type  should be manualy updated with the following SQL in the database;
+update snmp_metric_cfg set Conversion=3 where datasrctype='OCTETSTRING';
+* STRINGPARSER SnmpMetric Type no longer will be Scaled with Scale/shift values.
+
 
 # v 0.7.7 (2018-05-28)
 ### New Features